extrapolate WordOfTheDay into own repo
